apps/requirements.txt
anyio==3.7.0
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorlog==6.7.0
exceptiongroup==1.1.1
fastjsonschema==2.17.1
Flask-Cors==3.0.10
flask-sock==0.6.0
Flask==2.3.2
flask_login==0.6.2
flask_sqlalchemy==3.0.3
marshmallow-sqlalchemy
flask_migrate==4.0.4
flask_redis==0.4.0
flask_marshmallow
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
lxml==4.9.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
pip==23.1.2
reactpy==1.0.0
requests==2.31.0
setuptools==67.8.0
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
typing-extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
wheel==0.40.0
wsproto==1.2.0
zipp==3.15.0
python-dotenv

